+++ packages/DigitalScroll/index.js
@@ -1,7 +1,7 @@
 import DigitalScroll from './src/digital-scroll.vue';
 
 DigitalScroll.install = (App) => {
-	App.component(DigitalScroll.__name, DigitalScroll);
+	App.component(DigitalScroll.alanComponentName, DigitalScroll);
 };
 
 export default DigitalScroll;