+++ packages/index.js
@@ -1,16 +1,16 @@
 import StarrySky from './StarrySky';
 import DigitalScroll from './DigitalScroll';
-import Button from './Button';
+import AlanButton from './Button';
 import "./styles/index.scss"
 
 //按需引入
-export { StarrySky,DigitalScroll,Button };
+export { StarrySky,DigitalScroll,AlanButton };
 
-const components = [StarrySky,DigitalScroll,Button];
+const components = [StarrySky,DigitalScroll,AlanButton];
 
 const install = (App) => {
 	components.forEach((item) => {
-		App.component(item.__name, item);
+		App.component(item.alanComponentName, item);
 	});
 };