+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## <small>1.0.4 (2022-10-25)</small>
+## <small>1.0.11 (2022-10-25)</small>
 
 * ✨新增: index.html 格式调整 ([f769afa](https://github.com/2401345934/vue3-alan-vite-component/commit/f769afa))
 * ✨新增(custom): 修改 vite 配置 ([0ccb107](https://github.com/2401345934/vue3-alan-vite-component/commit/0ccb107))
@@ -7,6 +7,7 @@
 * ✨新增(package): 更新 md 说明 引入到 组件库地址 ([3218d51](https://github.com/2401345934/vue3-alan-vite-component/commit/3218d51))
 * ✨新增(package): 新增自定义说明提交 ([db8e26b](https://github.com/2401345934/vue3-alan-vite-component/commit/db8e26b))
 * ✨新增(package): 修改 md ([c609134](https://github.com/2401345934/vue3-alan-vite-component/commit/c609134))
+* ✨新增(package): 增加 Button 组件 ([7d191c1](https://github.com/2401345934/vue3-alan-vite-component/commit/7d191c1))
 * ✨新增(packages/DigitalScroll): 引入 ts ([9579d53](https://github.com/2401345934/vue3-alan-vite-component/commit/9579d53))
 * 0.1.3版本发布 ([b09708d](https://github.com/2401345934/vue3-alan-vite-component/commit/b09708d))
 * 更新md ([0c4ca78](https://github.com/2401345934/vue3-alan-vite-component/commit/0c4ca78))